TaijaBase

Documentation for TaijaBase.

TaijaBase.AbstractParallelizer

TaijaBase.Samplers.AbstractSampler

TaijaBase.Samplers.AbstractSampler

TaijaBase.Samplers.AbstractSamplingRule

TaijaBase.Samplers.ConditionalSampler

TaijaBase.Samplers.ConditionalSampler

TaijaBase.Samplers.ImproperSGLD

TaijaBase.Samplers.JointSampler

TaijaBase.Samplers.JointSampler

TaijaBase.Samplers.SGLD

TaijaBase.Samplers.UnconditionalSampler

TaijaBase.Samplers.UnconditionalSampler

TaijaBase.Samplers.PCD

TaijaBase.Samplers._energy

TaijaBase.Samplers._energy

TaijaBase.Samplers.energy

TaijaBase.Samplers.energy

TaijaBase.Samplers.energy

TaijaBase.Samplers.mcmc_samples

TaijaBase.Samplers.mcmc_samples

TaijaBase.Samplers.mcmc_samples

TaijaBase.parallelize

TaijaBase.AbstractParallelizer — Type

An abstract type for parallelizers. This type is owned by the TaijaBase module to allow shipping it to packages that depend on TaijaBase, without having to depend on the TaijaParallel module. See also parallelize for a detailed explanation.

TaijaBase.parallelize — Method

parallelize(
    parallelizer::nothing,
    f::Function,
    args...;
    kwargs...,
)

If no AbstractParallelizer has been supplied, just call or broadcast the function.

Note for developers

This function is owned by the TaijaBase module to allow shipping it to packages that depend on TaijaBase, without having to depend on the TaijaParallel module. For example, in CounterfactualExplanations we rely on this function to parallelize the computation of counterfactual explanations. Unless the user explicitly provides a parallelizer, the function will just call the function f or broadcast it over the collection of arguments. There is no need to depend on TaijaParallel for this simple operation.

TaijaBase.Samplers.AbstractSampler — Type

Base type for samplers.

TaijaBase.Samplers.AbstractSampler — Method

(sampler::AbstractSampler)(
    model,
    rule::AbstractSamplingRule;
    niter::Int = 100,
    clip_grads::Union{Nothing,AbstractFloat} = 1e-2,
    n_samples::Union{Nothing,Int} = nothing,
    kwargs...,
)

Base method for generating Monte Carlo samples for a given models, sampler and sampling rule.

TaijaBase.Samplers.AbstractSamplingRule — Type

Base type for sampling rules.

TaijaBase.Samplers.ConditionalSampler — Type

ConditionalSampler <: AbstractSampler

Generates conditional samples: $x \sim p(x|y).$

TaijaBase.Samplers.ConditionalSampler — Method

ConditionalSampler(
    𝒟x::Distribution, 𝒟y::Distribution;
    input_size::Dims, batch_size::Int,
    max_len::Int=10000, prob_buffer::AbstractFloat=0.95
)

Outer constructor for ConditionalSampler.

TaijaBase.Samplers.ImproperSGLD — Type

ImproperSGLD(α::Real=2.0, σ::Real=0.01)

Improper SGLD optimizer.

Examples

opt = ImproperSGLD()

TaijaBase.Samplers.JointSampler — Type

JointSampler <: AbstractSampler

Generates unconditional samples by drawing directly from joint distribution: $x \sim p(x, y).$

TaijaBase.Samplers.JointSampler — Method

JointSampler(
    𝒟x::Distribution, 𝒟y::Distribution, input_size::Dims, batch_size::Int;
    max_len::Int=10000, prob_buffer::AbstractFloat=0.95
)

Outer constructor for JointSampler.

TaijaBase.Samplers.SGLD — Type

SGLD(a::Real=1.0, b::Real=1.0, γ::Real=0.5)

Stochastic Gradient Langevin Dynamics (SGLD) optimizer.

Examples

opt = SGLD()
opt = SGLD(2.0, 100.0, 0.9)

TaijaBase.Samplers.UnconditionalSampler — Type

UnonditionalSampler <: AbstractSampler

Generates unconditional samples: $x \sim p(x).$

TaijaBase.Samplers.UnconditionalSampler — Method

UnconditionalSampler(
    𝒟x::Distribution,
    𝒟y::Union{Distribution,Nothing};
    input_size::Dims,
    batch_size::Int = 1,
    max_len::Int = 10000,
    prob_buffer::AbstractFloat = 0.95,
)

Outer constructor for UnonditionalSampler.

TaijaBase.Samplers.PCD — Method

PCD(
    sampler::AbstractSampler,
    model,
    rule::AbstractSamplingRule;
    ntransitions::Int = 100,
    niter::Int = 100,
    kwargs...,
)

Persistent Contrastive Divergence (PCD) algorithm. This algorithm was originally proposed by Tieleman (2008) and is a variant of the Contrastive Divergence (CD) algorithm. The main difference is that PCD uses a persistent chain to estimate the negative phase of the gradient. This is done by keeping the state of the Markov chain between iterations.

In our context, the sampler is the persistent chain and the model is a supervised model. The sampler generates samples from the model's learned distribution.

Note

This function does not perform any training. It only generates samples from the model. For training Joint Energy Models, see JointEnergyModels.jl.

Arguments

sampler::AbstractSampler: The sampler to use.

model: The model to sample from.

rule::AbstractSamplingRule: The sampling rule to use.

ntransitions::Int=100: The number of transitions to perform.

niter::Int=100: The number of iterations to perform.

kwargs...: Additional keyword arguments.

Returns

sampler.buffer: The buffer containing the samples generated by the sampler.

TaijaBase.Samplers._energy — Method

energy(f, x, y::Int; agg=mean)

Computes the energy for conditional samples $x \sim p_{\theta}(x|y)$: $E(x)=- f_{\theta}(x)[y]$.

TaijaBase.Samplers._energy — Method

energy(f, x)

Computes the energy for unconditional samples $x \sim p_{\theta}(x)$: $E(x)=-\text{LogSumExp}_y f_{\theta}(x)[y]$.

TaijaBase.Samplers.energy — Method

energy(sampler::ConditionalSampler, model, x, y)

Energy function for ConditionalSampler.

TaijaBase.Samplers.energy — Method

energy(sampler::JointSampler, model, x, y)

Energy function for JointSampler.

TaijaBase.Samplers.energy — Method

energy(sampler::UnconditionalSampler, model, x, y)

Energy function for UnconditionalSampler.

TaijaBase.Samplers.mcmc_samples — Method

mcmc_samples(
    sampler::ConditionalSampler,
    model,
    rule::Optimisers.AbstractRule,
    input_samples::AbstractArray;
    niter::Int,
    y::Union{Nothing,Int} = nothing,
)

Sampling method for ConditionalSampler.

TaijaBase.Samplers.mcmc_samples — Method

mcmc_samples(
    sampler::JointSampler,
    model,
    rule::Optimisers.AbstractRule,
    input_samples::AbstractArray;
    niter::Int,
    y::Union{Nothing,Int} = nothing,
)

Sampling method for JointSampler.

TaijaBase.Samplers.mcmc_samples — Method

mcmc_samples(
    sampler::UnconditionalSampler,
    model,
    rule::Optimisers.AbstractRule,
    input_samples::AbstractArray;
    niter::Int,
    y::Union{Nothing,Int} = nothing,
)

Sampling method for UnconditionalSampler.